# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Thrift Store",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/autoEcole": {
            "get": {
                "tags": [
                    "AutoEcole"
                ],
                "summary": "Récupère la liste de toutes les auto-écoles",
                "description": "Retourne une liste de toutes les auto-écoles disponibles",
                "operationId": "getAutoEcoles",
                "responses": {
                    "200": {
                        "description": "Liste des auto-écoles récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AutoEcole"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aucune auto-école n'a été trouvée"
                    },
                    "500": {
                        "description": "Erreur interne du serveur"
                    }
                }
            }
        },
        "/api/autoEcole/user": {
            "get": {
                "tags": [
                    "AutoEcole"
                ],
                "summary": "Obtient toutes les auto-écoles",
                "description": "Renvoie les auto ecole pour chaque utilisateur",
                "operationId": "getUserAutoEcole",
                "responses": {
                    "200": {
                        "description": "Liste des auto-écoles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AutoEcole"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur interne du serveur"
                    }
                }
            }
        },
        "/api/autoEcole/store": {
            "post": {
                "tags": [
                    "AutoEcole"
                ],
                "summary": "Crée une nouvelle auto-école",
                "description": "Crée une nouvelle auto-école avec les données fournies dans la requête",
                "operationId": "createAutoEcole",
                "requestBody": {
                    "description": "Données de l'auto-école à créer",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AutoEcole"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Nouvelle auto-école créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AutoEcole"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "L'utilisateur n'est pas un administrateur ou les données fournies sont invalides"
                    }
                }
            }
        },
        "/api/autoEcole/show/{id}": {
            "get": {
                "tags": [
                    "AutoEcole"
                ],
                "summary": "Récupère les détails d'une auto-école spécifique",
                "description": "Récupère les détails d'une auto-école en fonction de son ID",
                "operationId": "getAutoEcoleById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'auto-école à récupérer",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de l'auto-école récupérés avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AutoEcole"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "ID non trouvé, aucune auto-école correspondante",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "message": "Votre ID n'est pas trouvé"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/autoEcole/update/{id}": {
            "put": {
                "tags": [
                    "AutoEcole"
                ],
                "summary": "Mise à jour des détails d'une auto-école",
                "description": "Mise à jour des détails d'une auto-école existante en fonction de son ID",
                "operationId": "updateAutoEcole",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'auto-école à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Objet de mise à jour de l'auto-école",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AutoEcole"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Détails de l'auto-école mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AutoEcole"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "ID non trouvé, aucune auto-école correspondante",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "message": "Auto-école non trouvée"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/autoEcole/findAutoEcoleByUserId/{id}": {
            "get": {
                "tags": [
                    "AutoEcole"
                ],
                "summary": "Affiche les détails de l'auto-école associée à un utilisateur",
                "description": "Affiche les détails de l'auto-école associée à un utilisateur en fonction de son ID",
                "operationId": "showAutoEcoleByUserId",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de l'auto-école associée à l'utilisateur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "autoEcole": {
                                            "$ref": "#/components/schemas/AutoEcole"
                                        },
                                        "admin": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Utilisateur non trouvé, aucune auto-école associée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "message": "User has no associated autoEcole"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/autoEcole/delete/{id}": {
            "delete": {
                "tags": [
                    "AutoEcole"
                ],
                "summary": "Supprime une auto-école et l'utilisateur associé",
                "description": "Supprime une auto-école et l'utilisateur associé en fonction de l'ID de l'utilisateur",
                "operationId": "deleteAutoEcole",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Auto-école et utilisateur associé supprimés avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "example": "autoEcole and associated user deleted successfully"
                            }
                        }
                    },
                    "404": {
                        "description": "Auto-école non trouvée, aucun utilisateur associé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "message": "autoEcole not found"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Impossible de supprimer l'auto-école ou l'utilisateur associé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "message": "Failed to delete associated user"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "user_name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string"
                    },
                    "cin": {
                        "type": "string"
                    },
                    "numTel": {
                        "type": "string"
                    },
                    "dateNaissance": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "AutoEcole": {
                "title": "AutoEcole",
                "description": "Modèle de l'auto-école",
                "required": [
                    "nom",
                    "adresse"
                ],
                "properties": {
                    "nom": {
                        "description": "Nom de l'auto-école",
                        "type": "string"
                    },
                    "adresse": {
                        "description": "Adresse de l'auto-école",
                        "type": "string"
                    },
                    "autoecole_image": {
                        "description": "Image de l'auto-école",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}